module-service/requirements.txt
Flask==2.3.3
requests==2.31.0
pydantic==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23